--- a/packages/vscode-extension/src/builders/BuildManager.ts
+++ b/packages/vscode-extension/src/builders/BuildManager.ts
@@ -138,7 +138,7 @@ export class BuildManager {
         );
       }
 
-      await this.buildCache.storeBuild(buildFingerprint, buildResult, platform);
+      await this.buildCache.storeBuild(buildFingerprint, buildResult);
 
       return buildResult;
     };