--- a/packages/vscode-extension/src/builders/eas.ts
+++ b/packages/vscode-extension/src/builders/eas.ts
@@ -14,15 +14,16 @@ import { extractTarApp } from "./utils";
 export async function fetchEasBuild(
   cancelToken: CancelToken,
   config: EasConfig,
-  platform: DevicePlatform
+  platform: DevicePlatform,
+  appRoot: string
 ): Promise<string | undefined> {
-  if (!(await isEasCliInstalled())) {
+  if (!(await isEasCliInstalled(appRoot))) {
     throw new Error(
       "Failed to build iOS app using EAS build. Check if eas-cli is installed and available in PATH."
     );
   }
 
-  const build = await fetchBuild(config, platform);
+  const build = await fetchBuild(config, platform, appRoot);
   if (!build) {
     return undefined;
   }
@@ -36,9 +37,9 @@ export async function fetchEasBuild(
   return easBinaryPath;
 }
 
-async function fetchBuild(config: EasConfig, platform: DevicePlatform) {
+async function fetchBuild(config: EasConfig, platform: DevicePlatform, appRoot: string) {
   if (config.buildUUID) {
-    const build = await viewEasBuild(config.buildUUID, platform);
+    const build = await viewEasBuild(config.buildUUID, platform, appRoot);
     if (!build) {
       Logger.error(
         `Failed to find EAS build artifact with ID ${config.buildUUID} for platform ${platform}.`
@@ -54,7 +55,7 @@ async function fetchBuild(config: EasConfig, platform: DevicePlatform) {
     return build;
   }
 
-  const builds = await listEasBuilds(platform, config.profile);
+  const builds = await listEasBuilds(platform, config.profile, appRoot);
   if (!builds || builds.length === 0) {
     Logger.error(
       `Failed to find any EAS build artifacts for ${platform} with ${config.profile} profile. If you're building iOS app, make sure you set '"ios.simulator": true' option in eas.json.`