… packages/vscode-extension/src/panels/LaunchConfigController.ts
… packages/vscode-extension/src/panels/WebviewController.ts
… packages/vscode-extension/src/project/ApplicationContext.ts
… packages/vscode-extension/src/project/ApplicationContext1.ts
… packages/vscode-extension/src/project/applicationContext.ts
… packages/vscode-extension/src/project/deviceSession.ts
… packages/vscode-extension/src/project/ide.ts
… packages/vscode-extension/src/project/metro.ts
… packages/vscode-extension/src/project/project.ts
… packages/vscode-extension/src/utilities/common.ts
… packages/vscode-extension/src/utilities/expoCli.ts
… packages/vscode-extension/src/utilities/findAndSetupNewAppRootFolder.ts
… packages/vscode-extension/src/utilities/packageManager.ts
… packages/vscode-extension/src/utilities/reactNative.ts
… packages/vscode-extension/src/webview/providers/LaunchConfigProvider.tsx
… packages/vscode-extension/src/webview/views/LaunchConfigurationView.css
… packages/vscode-extension/src/webview/views/LaunchConfigurationView.tsx